backend/requirements.txt
# Flask and Core Extensions
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Executor==1.0.0
Werkzeug==3.0.1

# Authentication & Authorization
Flask-JWT-Extended==4.6.0
PyJWT==2.8.0
bcrypt==4.1.2

# WSGI Server
gunicorn==21.2.0

# Database Driver
psycopg2-binary==2.9.9

# HTTP Clients
requests==2.31.0
httpx==0.27.0

# OpenAI / Azure OpenAI
openai>=1.12.0

# Azure SDKs
azure-cognitiveservices-speech==1.34.0
azure-identity==1.15.0
azure-core==1.30.0
azure-storage-blob==12.19.1
azure-storage-file-share==12.15.0

# BotBuilder (if still actively used)
# botbuilder-core==4.16.2
# botbuilder-integration-aiohttp==4.16.2
# Commented out as Teams routes look incomplete/placeholder

# Utilities
python-dotenv==0.21.0
python-dateutil==2.8.2

# Optional
# Pillow==9.3.0